TSL Testing: split the tests into separate files
